move otel env functions to otel package
