+++ packages/zero-cache/src/server/otel-start.ts
@@ -20,6 +20,12 @@ import {
 import {PeriodicExportingMetricReader} from '@opentelemetry/sdk-metrics';
 import {NodeSDK} from '@opentelemetry/sdk-node';
 import {ATTR_SERVICE_VERSION} from '@opentelemetry/semantic-conventions';
+import {
+  otelEnabled,
+  otelLogsEnabled,
+  otelMetricsEnabled,
+  otelTracesEnabled,
+} from '../../../otel/src/enabled.ts';
 
 class OtelManager {
   static #instance: OtelManager;
@@ -102,31 +108,6 @@ class OtelManager {
       body: 'OpenTelemetry SDK started successfully',
     });
   }
-}
-
-export function otelEnabled() {
-  return otelMetricsEnabled() || otelTracesEnabled() || otelLogsEnabled();
-}
-
-export function otelMetricsEnabled() {
-  return (
-    process.env.OTEL_EXPORTER_OTLP_ENDPOINT ||
-    process.env.OTEL_EXPORTER_OTLP_METRICS_ENDPOINT
-  );
-}
-
-export function otelLogsEnabled() {
-  return (
-    process.env.OTEL_EXPORTER_OTLP_ENDPOINT ||
-    process.env.OTEL_EXPORTER_OTLP_LOGS_ENDPOINT
-  );
-}
-
-export function otelTracesEnabled() {
-  return (
-    process.env.OTEL_EXPORTER_OTLP_ENDPOINT ||
-    process.env.OTEL_EXPORTER_OTLP_TRACES_ENDPOINT
-  );
 }
 
 export const startOtelAuto = () => OtelManager.getInstance().startOtelAuto();
